requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
class==1.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.1.1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
Django==3.1.7
dnspython==2.0.0
docopt==0.6.2
docutils==0.16
dropbox==11.2.0
email-validator==1.1.2
et-xmlfile==1.0.1
filelock==3.0.12
firebase-admin==4.4.0
future==0.18.2
gast==0.3.3
gcloud==0.18.3
google==3.0.0
google-api-core==1.22.4
google-api-python-client==1.12.4
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
httplib2==0.18.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imutils==0.5.3
ipython==7.18.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
joblib==0.17.0
jwcrypto==0.8
katna==0.4.1.4
Keras==2.4.3
Keras-Preprocessing==1.1.2
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kivymd==0.104.1
kiwisolver==1.2.0
Markdown==3.3.2
matplotlib==3.3.2
more-itertools==8.7.0
moviepy==1.0.3
msgpack==1.0.0
networkx==2.5
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
openpyxl==3.0.5
opt-einsum==3.3.0
pandas==1.1.3
parso==0.7.1
password-validator==1.0
pefile==2019.4.18
phonenumbers==8.12.15
pickleshare==0.7.5
Pillow==7.2.0
pipreqs==0.4.10
ply==3.11
postcode-validator-uk==0.0.5
proglog==0.1.9
prompt-toolkit==3.0.8
protobuf==3.13.0
py3-validate-email==0.2.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
Pygments==2.7.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.15.1.1.7.5
pyrebase5==5.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-jwt==3.3.0
python-resize-image==1.1.19
pytz==2020.1
PyWavelets==1.1.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.3
six==1.15.0
smartcrop==0.3.2
soupsieve==2.0.1
sqlparse==0.4.1
sseclient==0.0.27
stone==3.2.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.10.1
tqdm==4.50.2
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.4.2
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
yarg==0.1.9
yolk==0.4.3
